dynamic list
ability to load initial list

remove comments

pass in any object

remove unecessary push

initial list now mandatory

use listOrConfig

rename to fieldFactory

added hooks and util function

remove state comment

wording for action

wording for action

added changelog and doc

added tests

remove comment

fix lint

lint fix

added baselist

update documentation

move reducer to hooks folder

potentially fix some lint

update action names

fix lint and updated comments
