+++ packages/react-form/src/hooks/list/index.ts
@@ -1,1 +1,2 @@
 export {useList} from './list';
+export {useDynamicList} from './dynamiclist';
