+++ packages/react-form/src/hooks/list/list.ts
@@ -1,35 +1,9 @@
-import {useMemo, useEffect, ChangeEvent} from 'react';
-import isEqual from 'fast-deep-equal';
+import {FieldDictionary} from '../../types';
 
-import {
-  ValidationDictionary,
-  Validator,
-  NormalizedValidationDictionary,
-  FieldStates,
-  FieldDictionary,
-  FieldState,
-  ErrorValue,
-  ListValidationContext,
-} from '../../types';
-import {mapObject, normalizeValidation, isChangeEvent} from '../../utilities';
-
-import {
-  updateAction,
-  updateErrorAction,
-  newDefaultAction,
-  reinitializeAction,
-  resetAction,
-  useListReducer,
-} from './reducer';
-
-export interface FieldListConfig<Item extends object> {
-  list: Item[];
-  validates?: Partial<ValidationDictionary<Item, ListValidationContext<Item>>>;
-}
+import {useBaseList, FieldListConfig} from './baselist';
 
 /**
- * A custom hook for handling the state and validations of fields for a list of objects.
- *
+ * A custom hook for utilizing useBaseList. useList uses useBaseList. This hook is responsible for returning just the field objects useBaseList provides.
  * In it's simplest form `useList` can be called with a single parameter with the list to derive default values and structure from.
  *
  * ```typescript
@@ -118,154 +92,16 @@ export interface FieldListConfig<Item extends object> {
  * }
  * ```
  *
- * @param config - A configuration object specifying both the value and validation config.
+ * @param listOrConfig - A configuration object specifying both the value and validation config.
  * @param validationDependencies - An array of dependencies to use to decide when to regenerate validators.
  * @returns A list of dictionaries of `Field` objects representing the state of your input. It also includes functions to manipulate that state. Generally, you will want to pass these callbacks down to the component or components representing your input.
  *
- * @remarks
- * **Reinitialization:** If the `list` property of the field configuration changes between calls to `useList`,
- * the field will be reset to use it as it's new default value.
- *
- * **Imperative methods:** The returned `Field` objects contains a number of methods used to imperatively alter their state.
- * These should only be used as escape hatches where the existing hooks and components do not make your life easy,
- * or to build new abstractions in the same vein as `useForm`, `useSubmit` and friends.
  */
 export function useList<Item extends object>(
   listOrConfig: FieldListConfig<Item> | Item[],
   validationDependencies: unknown[] = [],
 ): FieldDictionary<Item>[] {
-  const list = Array.isArray(listOrConfig) ? listOrConfig : listOrConfig.list;
-  const validates: FieldListConfig<Item>['validates'] = Array.isArray(
-    listOrConfig,
-  )
-    ? {}
-    : listOrConfig.validates || {};
+  const {fields} = useBaseList(listOrConfig, validationDependencies);
 
-  const [state, dispatch] = useListReducer(list);
-
-  useEffect(() => {
-    if (!isEqual(list, state.initial)) {
-      dispatch(reinitializeAction(list));
-    }
-  }, [list, state.initial, dispatch]);
-
-  const validationConfigs = useMemo(
-    () =>
-      mapObject<NormalizedValidationDictionary<any>>(
-        validates,
-        normalizeValidation,
-      ),
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-    [validates, ...validationDependencies],
-  );
-
-  const handlers = useMemo(() => {
-    return state.list.map((item, index) => {
-      return mapObject<FieldDictionary<Item>>(
-        item,
-        <Key extends keyof Item & string>(
-          field: FieldState<Item[Key]>,
-          key: Key,
-        ) => {
-          const target = {index, key};
-
-          function validate(value = field.value) {
-            const validates = validationConfigs[key];
-
-            if (validates == null) {
-              return;
-            }
-
-            const siblings = state.list.filter(listItem => listItem !== item);
-
-            return runValidation(
-              error =>
-                dispatch(
-                  updateErrorAction<Item>({target, error: error || ''}),
-                ),
-              {value, siblings, listItem: item},
-              validates,
-            );
-          }
-
-          return {
-            onChange(value: Item[Key] | ChangeEvent) {
-              const normalizedValue = (isChangeEvent(value)
-                ? value.target.value
-                : value) as Item[Key];
-
-              dispatch(
-                updateAction({
-                  target,
-                  value: normalizedValue,
-                }),
-              );
-
-              if (field.error) {
-                validate(normalizedValue);
-              }
-            },
-            reset() {
-              dispatch(resetAction({target}));
-            },
-            newDefaultValue(value: Item[Key]) {
-              dispatch(newDefaultAction({target, value}));
-            },
-            runValidation: validate,
-            onBlur() {
-              const {touched, error} = field;
-
-              if (touched === false && error == null) {
-                return;
-              }
-              validate();
-            },
-            setError(error: string) {
-              dispatch(updateErrorAction({target, error}));
-            },
-          };
-        },
-      );
-    });
-  }, [dispatch, state.list, validationConfigs]);
-
-  return useMemo(() => {
-    return state.list.map((item, index) => {
-      return mapObject(item, (field, key: keyof Item) => {
-        return {
-          ...field,
-          ...(handlers[index][key] as any),
-        };
-      });
-    });
-  }, [state.list, handlers]);
-}
-
-function runValidation<Value, Record extends object>(
-  updateError: (error: ErrorValue) => void,
-  state: {
-    value: Value;
-    listItem: FieldStates<Record>;
-    siblings: FieldStates<Record>[];
-  },
-  validators: Validator<Value, ListValidationContext<Record>>[],
-) {
-  const {value, listItem, siblings} = state;
-
-  const error = validators
-    .map(check =>
-      check(value, {
-        listItem,
-        siblings,
-      }),
-    )
-    .filter(value => value != null);
-
-  if (error && error.length > 0) {
-    const [firstError] = error;
-    updateError(firstError);
-    return firstError;
-  }
-
-  updateError(undefined);
+  return fields;
 }
