+++ packages/react-form/src/hooks/list/utils/index.ts
@@ -1,0 +1,1 @@
+export {runValidation} from './utils';
