+++ packages/react-form/src/index.ts
@@ -14,6 +14,7 @@ export {
   useForm,
   submitFail,
   submitSuccess,
+  useDynamicList,
 } from './hooks';
 
 export {
